sua sai so khi tinh Calories can nap

+++ HealthManagementTraining/src/main/java/com/sixthgroup/healthmanagementtraining/services/AdminFoodServices.java
@@ -85,7 +85,7 @@ public class AdminFoodServices {
         try (Connection conn = JdbcUtils.getConn()) {
             PreparedStatement stm = conn.prepareStatement("INSERT INTO food (foodName, caloriesPerUnit, lipidPerUnit, proteinPerUnit, fiberPerUnit, foodCategory_id, unitType) VALUES (?, ?, ?, ?, ?, ?, ?)");
             stm.setString(1, food.getFoodName());
-            stm.setInt(2, food.getCaloriesPerUnit());
+            stm.setFloat(2, food.getCaloriesPerUnit());
             stm.setFloat(3, food.getLipidPerUnit());
             stm.setFloat(4, food.getProteinPerUnit());
             stm.setFloat(5, food.getFiberPerUnit());
@@ -112,7 +112,7 @@ public class AdminFoodServices {
             String sql = "UPDATE food SET foodName = ?, caloriesPerUnit = ?, lipidPerUnit = ?, proteinPerUnit = ?, fiberPerUnit = ? WHERE id = ?";
             PreparedStatement stm = conn.prepareStatement(sql);
             stm.setString(1, food.getFoodName());
-            stm.setInt(2, food.getCaloriesPerUnit());
+            stm.setFloat(2, food.getCaloriesPerUnit());
             stm.setFloat(3, food.getLipidPerUnit());
             stm.setFloat(4, food.getProteinPerUnit());
             stm.setFloat(5, food.getFiberPerUnit());
